atc: behaviour: support queueing notification bus
...and support receiving string payloads from the bus

The current behaviour of the notification bus when a new notification
comes in is:

Try to send that notification to each Go chan:
  * If that chan is empty, send along the chan
  * If that chan already has a notification waiting to be received, do
    nothing

This is nice because it avoids blocking. However, there are use cases
(like using the notification bus to send the ID of table entities that
changed) where this isn't desirable - even if the consumer is slow, we
want it to eventually receive every event.

What I really wanted to implement this was an unbounded channel, but
that's not in the cards obviously. Instead, associate each consumer with
a `queue` (implemented as a slice that can grow without bound + a
goroutine that drains the slice into the chan when it changes). That
way, we'll still never block when an event comes in, but we can ensure
we won't lose any messages even if the consumer is super slow.

Note that I tagged this commit as behaviour (since it introduces new
behaviour that can be used), but it doesn't yet change anything since
nothing is making use of it - everything is still using the "don't
queue notifications" behaviour, and nothing sends any payloads.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>