web: behaviour: section headers on dashboard
concourse/concourse#5439

We need to compute the offset for rendering the pipeline cards based on
the height of the section headers and the separator line. This is to
determine what is/isn't on screen for lazy rendering.

Also, we provide a default viewport size for the `whenOnDashboard`
helper to avoid running into weird "no pipelines set" issues in tests
due to lazy rendering.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Zoe Tian <ztian@pivotal.io>